extracting reverse diagram analytics & preparing cat c test report components
