+++ src/pages/page-names.constants.ts
@@ -47,20 +47,20 @@ export const CAT_BE = {
 };
 
 export const CAT_C = {
-  BACK_TO_OFFICE_PAGE: 'BackToOfficeCatBEPage',
-  COMMUNICATION_PAGE: 'CommunicationCatBEPage',
-  DEBRIEF_PAGE: 'DebriefCatBEPage',
-  HEALTH_DECLARATION_PAGE: 'HealthDeclarationCatBEPage',
-  OFFICE_PAGE: 'OfficeCatBEPage',
-  PASS_FINALISATION_PAGE: 'PassFinalisationCatBEPage',
-  REVERSE_DIAGRAM_PAGE: 'ReverseDiagramCatBEPage',
-  TEST_REPORT_PAGE: 'TestReportCatBEPage',
-  WAITING_ROOM_PAGE: 'WaitingRoomCatBEPage',
-  WAITING_ROOM_TO_CAR_PAGE: 'WaitingRoomToCarCatBEPage',
-  POST_DEBRIEF_HOLDING_PAGE: 'PostDebriefHoldingCatBEPage',
-  REKEY_REASON_PAGE: 'RekeyReasonCatBEPage',
+  BACK_TO_OFFICE_PAGE: 'BackToOfficeCatCPage',
+  COMMUNICATION_PAGE: 'CommunicationCatCPage',
+  DEBRIEF_PAGE: 'DebriefCatCPage',
+  HEALTH_DECLARATION_PAGE: 'HealthDeclarationCatCPage',
+  OFFICE_PAGE: 'OfficeCatCPage',
+  PASS_FINALISATION_PAGE: 'PassFinalisationCatCPage',
+  REVERSE_DIAGRAM_PAGE: 'ReverseDiagramCatCPage',
+  TEST_REPORT_PAGE: 'TestReportCatCPage',
+  WAITING_ROOM_PAGE: 'WaitingRoomCatCPage',
+  WAITING_ROOM_TO_CAR_PAGE: 'WaitingRoomToCarCatCPage',
+  POST_DEBRIEF_HOLDING_PAGE: 'PostDebriefHoldingCatCPage',
+  REKEY_REASON_PAGE: 'RekeyReasonCatCPage',
   REKEY_UPLOAD_OUTCOME_PAGE: 'RekeyUploadOutcomePage',
-  NON_PASS_FINALISATION_PAGE: 'NonPassFinalisationCatBEPage',
-  VEHICLE_CHECKS_MODAL: 'VehicleChecksCatBEModal',
-  VIEW_TEST_RESULT_PAGE:'ViewTestResultCatBEPage',
+  NON_PASS_FINALISATION_PAGE: 'NonPassFinalisationCatCPage',
+  VEHICLE_CHECKS_MODAL: 'VehicleChecksCatCModal',
+  VIEW_TEST_RESULT_PAGE:'ViewTestResultCatCPage',
 };
