+++ src/pages/test-report/cat-c/components/reverse-diagram-modal.cat-c/reverse-diagram-modal.cat-c.scss
@@ -1,4 +1,4 @@
-reverse-diagram-modal {
+reverse-diagram-modal-cat-c {
   .exclamation-indicator {
     width: 26px;
     height: 26px;
