+++ src/pages/test-report/components/reverse-diagram-modal/__tests__/reverse-diagram-modal.analytics.effects.spec.ts
@@ -4,27 +4,27 @@ import { ReplaySubject } from 'rxjs/ReplaySubject';
 import { StoreModule, Store } from '@ngrx/store';
 import { provideMockActions } from '@ngrx/effects/testing';
 import * as reverseDiagramModalActions from '../reverse-diagram-modal.actions';
-import { AnalyticsProvider } from '../../../../../../providers/analytics/analytics';
-import { AnalyticsProviderMock } from '../../../../../../providers/analytics/__mocks__/analytics.mock';
+import { AnalyticsProvider } from '../../../../../providers/analytics/analytics';
+import { AnalyticsProviderMock } from '../../../../../providers/analytics/__mocks__/analytics.mock';
 import {
   AnalyticsDimensionIndices,
   AnalyticsScreenNames,
   AnalyticsEventCategories,
   AnalyticsEvents,
-} from '../../../../../../providers/analytics/analytics.model';
-import { AnalyticRecorded } from '../../../../../../providers/analytics/analytics.actions';
-import { StoreModel } from '../../../../../../shared/models/store.model';
-import * as testsActions from '../../../../../../modules/tests/tests.actions';
-import * as fakeJournalActions from '../../../../../fake-journal/fake-journal.actions';
-import { testsReducer } from '../../../../../../modules/tests/tests.reducer';
-import { testReportPracticeModeSlot, candidateMock } from '../../../../../../modules/tests/__mocks__/tests.mock';
-import { PopulateCandidateDetails } from '../../../../../../modules/tests/journal-data/candidate/candidate.actions';
+} from '../../../../../providers/analytics/analytics.model';
+import { AnalyticRecorded } from '../../../../../providers/analytics/analytics.actions';
+import { StoreModel } from '../../../../../shared/models/store.model';
+import * as testsActions from '../../../../../modules/tests/tests.actions';
+import * as fakeJournalActions from '../../../../fake-journal/fake-journal.actions';
+import { testsReducer } from '../../../../../modules/tests/tests.reducer';
+import { testReportPracticeModeSlot, candidateMock } from '../../../../../modules/tests/__mocks__/tests.mock';
+import { PopulateCandidateDetails } from '../../../../../modules/tests/journal-data/candidate/candidate.actions';
 import { Application } from '@dvsa/mes-journal-schema';
-import { end2endPracticeSlotId } from '../../../../../../shared/mocks/test-slot-ids.mock';
+import { end2endPracticeSlotId } from '../../../../../shared/mocks/test-slot-ids.mock';
 import * as applicationReferenceActions
-  from '../../../../../../modules/tests/journal-data/application-reference/application-reference.actions';
+  from '../../../../../modules/tests/journal-data/application-reference/application-reference.actions';
 import { TestCategory } from '@dvsa/mes-test-schema/categories/common/test-category';
-import { PopulateTestCategory } from '../../../../../../modules/tests/category/category.actions';
+import { PopulateTestCategory } from '../../../../../modules/tests/category/category.actions';
 
 describe('Reverse Diagram Modal Analytics Effects', () => {
 
@@ -63,9 +63,12 @@ describe('Reverse Diagram Modal Analytics Effects', () => {
   describe('reverseDiagramViewDidEnter', () => {
     it('should call setCurrentPage and addCustomDimension', (done) => {
       // ARRANGE
+
+      // TODO: MES-4287 Change the category to C
       store$.dispatch(new testsActions.StartTest(123, TestCategory.BE));
       store$.dispatch(new PopulateCandidateDetails(candidateMock));
       store$.dispatch(new applicationReferenceActions.PopulateApplicationReference(mockApplication));
+      // TODO: MES-4287 Change the category to C
       store$.dispatch(new PopulateTestCategory(TestCategory.BE));
       // ACT
       actions$.next(new reverseDiagramModalActions.ReverseDiagramViewDidEnter());
@@ -73,7 +76,9 @@ describe('Reverse Diagram Modal Analytics Effects', () => {
       effects.reverseDiagramViewDidEnter$.subscribe((result) => {
         expect(result instanceof AnalyticRecorded).toBe(true);
         expect(analyticsProviderMock.addCustomDimension)
-        .toHaveBeenCalledWith(AnalyticsDimensionIndices.TEST_CATEGORY, 'B+E');
+
+          // TODO: MES-4287 Change the category to C
+          .toHaveBeenCalledWith(AnalyticsDimensionIndices.TEST_CATEGORY, 'B+E');
         expect(analyticsProviderMock.addCustomDimension)
           .toHaveBeenCalledWith(AnalyticsDimensionIndices.CANDIDATE_ID, '1');
         expect(analyticsProviderMock.addCustomDimension)
@@ -95,7 +100,9 @@ describe('Reverse Diagram Modal Analytics Effects', () => {
       effects.reverseDiagramViewDidEnter$.subscribe((result) => {
         expect(result instanceof AnalyticRecorded).toBe(true);
         expect(analyticsProviderMock.addCustomDimension)
-        .toHaveBeenCalledWith(AnalyticsDimensionIndices.TEST_CATEGORY, 'B+E');
+
+          // TODO: MES-4287 Change the category to C
+          .toHaveBeenCalledWith(AnalyticsDimensionIndices.TEST_CATEGORY, 'B+E');
         expect(analyticsProviderMock.addCustomDimension)
           .toHaveBeenCalledWith(AnalyticsDimensionIndices.CANDIDATE_ID, '1');
         expect(analyticsProviderMock.addCustomDimension)
